refactor: 관리자 예약 및 시간 관리 API 분리 및 패키지 리팩토링
- 관리자 전용 예약 및 시간 관련 API를 별도 컨트롤러(AdminReservationWaitController, AdminReservationTimeController)로 분리
- 관련된 클래스 및 요청/응답 DTO를 admin.reservation 패키지로 이동
- 기존 컨트롤러(ReservationWaitController, ReservationTimeController)에서 관리자 기능 제거
- ReservationRepository에 @Repository 어노테이션 추가
