+++ src/main/java/roomescape/admin/reservation/AdminReservationController.java
@@ -1,4 +1,4 @@
-package roomescape.admin;
+package roomescape.admin.reservation;
 
 import jakarta.validation.Valid;
 import java.net.URI;
