+++ src/main/java/roomescape/admin/reservation/time/AdminReservationTimeController.java
@@ -1,0 +1,54 @@
+package roomescape.admin.reservation.time;
+
+import java.net.URI;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
+import roomescape.global.auth.Auth;
+import roomescape.member.domain.Role;
+import roomescape.reservation.time.application.ReservationTimeService;
+import roomescape.reservation.time.presentation.dto.ReservationTimeRequest;
+import roomescape.reservation.time.presentation.dto.ReservationTimeResponse;
+
+@RestController
+@RequestMapping("/times")
+public class AdminReservationTimeController {
+
+    private final ReservationTimeService reservationTimeService;
+
+    public AdminReservationTimeController(final ReservationTimeService reservationTimeService) {
+        this.reservationTimeService = reservationTimeService;
+    }
+
+    @Auth(Role.ADMIN)
+    @PostMapping
+    public ResponseEntity<ReservationTimeResponse> createTime(
+            final @RequestBody ReservationTimeRequest reservationTimeRequest
+    ) {
+        ReservationTimeResponse reservationTime = reservationTimeService.createReservationTime(reservationTimeRequest);
+
+        return ResponseEntity.created(createUri(reservationTime.getId()))
+                .body(reservationTime);
+    }
+
+    @Auth(Role.ADMIN)
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteReservationTime(
+            final @PathVariable Long id
+    ) {
+        reservationTimeService.deleteReservationTime(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    private URI createUri(Long timeId) {
+        return ServletUriComponentsBuilder.fromCurrentRequest()
+                .path("/{id}")
+                .buildAndExpand(timeId)
+                .toUri();
+    }
+}
