+++ src/main/java/roomescape/reservation/time/presentation/ReservationTimeController.java
@@ -1,23 +1,16 @@
 package roomescape.reservation.time.presentation;
 
-import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 import roomescape.global.auth.Auth;
 import roomescape.member.domain.Role;
 import roomescape.reservation.time.application.ReservationTimeService;
 import roomescape.reservation.time.presentation.dto.AvailableReservationTimeResponse;
-import roomescape.reservation.time.presentation.dto.ReservationTimeRequest;
 import roomescape.reservation.time.presentation.dto.ReservationTimeResponse;
 
 @RestController
@@ -28,17 +21,6 @@ public class ReservationTimeController {
 
     public ReservationTimeController(final ReservationTimeService reservationTimeService) {
         this.reservationTimeService = reservationTimeService;
-    }
-
-    @Auth(Role.ADMIN)
-    @PostMapping
-    public ResponseEntity<ReservationTimeResponse> createTime(
-            final @RequestBody ReservationTimeRequest reservationTimeRequest
-    ) {
-        ReservationTimeResponse reservationTime = reservationTimeService.createReservationTime(reservationTimeRequest);
-
-        return ResponseEntity.created(createUri(reservationTime.getId()))
-                .body(reservationTime);
     }
 
     @Auth(Role.USER)
@@ -59,21 +41,5 @@ public class ReservationTimeController {
         return ResponseEntity.ok().body(
                 reservationTimeService.getReservationTimes(date, themeId)
         );
-    }
-
-    @Auth(Role.ADMIN)
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservationTime(
-            final @PathVariable Long id
-    ) {
-        reservationTimeService.deleteReservationTime(id);
-        return ResponseEntity.noContent().build();
-    }
-
-    private URI createUri(Long timeId) {
-        return ServletUriComponentsBuilder.fromCurrentRequest()
-                .path("/{id}")
-                .buildAndExpand(timeId)
-                .toUri();
     }
 }
